requirements-pip.txt
Celery==5.4.0
django-admin-sortable2==2.2.1
django-appconf==1.0.6
django-crequest==2018.5.11
django-dirtyfields==1.9.5
django-galleryfield==2.2.0
django-ipware==7.0.1
django-multiselectfield==0.1.13
django-object-actions==4.2.0
django-sortedm2m==3.1.1
django-suit-v2-pm==2.4
django-tinymce==4.1.0
django-treenode==0.22.2
dj-database-url==2.2.0
easy_cache==2.0.0
filetype==1.2.0
git+https://github.com/etianen/django-watson.git@refs/pull/309/head
ipinfo==5.1.1
pytils==0.4.1
redis==5.2.1
regex==2024.11.6
